Extract classes from AlbumListViewModel
