+++ feature/album/src/main/kotlin/com/igorwojda/showcase/feature/album/presentation/screen/albumlist/AlbumListViewModel.kt
@@ -1,18 +1,9 @@
 package com.igorwojda.showcase.feature.album.presentation.screen.albumlist
 
-import androidx.compose.runtime.Immutable
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.viewModelScope
-import com.igorwojda.showcase.feature.album.domain.model.Album
 import com.igorwojda.showcase.feature.album.domain.usecase.GetAlbumListUseCase
-import com.igorwojda.showcase.feature.album.presentation.screen.albumlist.AlbumListViewModel.Action
-import com.igorwojda.showcase.feature.album.presentation.screen.albumlist.AlbumListViewModel.AlbumListUiState
-import com.igorwojda.showcase.feature.album.presentation.screen.albumlist.AlbumListViewModel.AlbumListUiState.Content
-import com.igorwojda.showcase.feature.album.presentation.screen.albumlist.AlbumListViewModel.AlbumListUiState.Error
-import com.igorwojda.showcase.feature.album.presentation.screen.albumlist.AlbumListViewModel.AlbumListUiState.Loading
 import com.igorwojda.showcase.feature.base.domain.result.Result
-import com.igorwojda.showcase.feature.base.presentation.viewmodel.BaseAction
-import com.igorwojda.showcase.feature.base.presentation.viewmodel.BaseState
 import com.igorwojda.showcase.feature.base.presentation.viewmodel.BaseViewModel
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.launch
@@ -20,7 +11,7 @@ import kotlinx.coroutines.launch
 internal class AlbumListViewModel(
     private val savedStateHandle: SavedStateHandle,
     private val getAlbumListUseCase: GetAlbumListUseCase,
-) : BaseViewModel<AlbumListUiState, Action>(Loading) {
+) : BaseViewModel<AlbumListUiState, AlbumListAction>(AlbumListUiState.Loading) {
     
     companion object {
         const val DEFAULT_QUERY_NAME = "Jackson"
@@ -41,53 +32,23 @@ internal class AlbumListViewModel(
 
         savedStateHandle[SAVED_QUERY_KEY] = query
 
-        sendAction(Action.AlbumListLoadStart)
+        sendAction(AlbumListAction.AlbumListLoadStart)
 
         job =
             viewModelScope.launch {
                 getAlbumListUseCase(query).also { result ->
-                    val action =
+                    val albumListAction =
                         when (result) {
                             is Result.Success -> {
-                                Action.AlbumListLoadSuccess(result.value)
+                                AlbumListAction.AlbumListLoadSuccess(result.value)
                             }
                             is Result.Failure -> {
-                                Action.AlbumListLoadFailure
+                                AlbumListAction.AlbumListLoadFailure
                             }
                         }
 
-                    sendAction(action)
+                    sendAction(albumListAction)
                 }
             }
-    }
-
-    internal sealed interface Action : BaseAction<AlbumListUiState> {
-        object AlbumListLoadStart : Action {
-            override fun reduce(state: AlbumListUiState) = Loading
-        }
-
-        class AlbumListLoadSuccess(
-            private val albums: List<Album>,
-        ) : Action {
-            override fun reduce(state: AlbumListUiState) = Content(albums)
-        }
-
-        object AlbumListLoadFailure : Action {
-            override fun reduce(state: AlbumListUiState) = Error
-        }
-    }
-
-    @Immutable
-    internal sealed interface AlbumListUiState : BaseState {
-        @Immutable
-        data class Content(
-            val albums: List<Album>,
-        ) : AlbumListUiState
-
-        @Immutable
-        data object Loading : AlbumListUiState
-
-        @Immutable
-        data object Error : AlbumListUiState
     }
 }
